pyproject.toml
[project]
name = "coastal-hotel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bertopic>=0.16",
    "hdbscan>=0.8.38",
    "jupyter>=1.1.1",
    "matplotlib>=3.9",
    "nbformat>=5.10.4",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pyvi>=0.1.1",
    "qdrant-client>=1.9",
    "scikit-learn>=1.5",
    "sentence-transformers>=3.0",
    "spacy>=3.8.13",
    "stopwords>=1.0.2",
    "stopwordsiso>=0.6.1",
    "tqdm>=4.67.3",
    "umap-learn>=0.5",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]
